Probability metrics for sf objects

tidysdm provides specialised metrics for SDMs, which have their own help pages(boyce_cont(), kap_max(), and tss_max()). Additionally, it also provides methods to handle sf::sf objects for the following standard yardstick metrics:

yardstick::average_precision()

yardstick::brier_class()

yardstick::classification_cost()

yardstick::gain_capture()

yardstick::mn_log_loss()

yardstick::pr_auc()

yardstick::roc_auc()

yardstick::roc_aunp()

yardstick::roc_aunu()

Usage

# S3 method for class 'sf'
average_precision(data, ...)

# S3 method for class 'sf'
brier_class(data, ...)

# S3 method for class 'sf'
classification_cost(data, ...)

# S3 method for class 'sf'
gain_capture(data, ...)

# S3 method for class 'sf'
mn_log_loss(data, ...)

# S3 method for class 'sf'
pr_auc(data, ...)

# S3 method for class 'sf'
roc_auc(data, ...)

# S3 method for class 'sf'
roc_aunp(data, ...)

# S3 method for class 'sf'
roc_aunu(data, ...)

Arguments

data

an sf::sf object

...

any other parameters to pass to the data.frame version of the metric. See the specific man page for the metric of interest.

Value

A tibble with columns .metric, .estimator, and .estimate and 1 row of values.

Details

Note that roc_aunp and roc_aunu are multiclass metrics, and as such are are not relevant for SDMs (which work on a binary response). They are included for completeness, so that all class probability metrics from yardstick have an sf method, for applications other than SDMs.