Obtain and format results produced by tuning functions for ensemble objects

Return a tibble of performance metrics for all models.

Usage

# S3 method for simple_ensemble
collect_metrics(x, ...)

# S3 method for repeat_ensemble
collect_metrics(x, ...)

Arguments

x

A simple_ensemble or repeat_ensemble object

...

Not currently used.

Value

A tibble.

Details

When applied to a ensemble, the metrics that are returned do not contain the actual tuning parameter columns and values (unlike when these collect functions are run on other objects). The reason is that ensembles contain different types of models or models with different tuning parameters.

See also

Examples

collect_metrics(lacerta_ensemble)
#> # A tibble: 12 × 5
#>    wflow_id       .metric     mean std_err     n
#>    <chr>          <chr>      <dbl>   <dbl> <int>
#>  1 default_glm    boyce_cont 0.633 0.133       5
#>  2 default_glm    roc_auc    0.963 0.00580     5
#>  3 default_glm    tss_max    0.866 0.0257      5
#>  4 default_rf     boyce_cont 0.596 0.133       5
#>  5 default_rf     roc_auc    0.980 0.00684     5
#>  6 default_rf     tss_max    0.899 0.0256      5
#>  7 default_gbm    boyce_cont 0.789 0.0906      5
#>  8 default_gbm    roc_auc    0.975 0.0105      5
#>  9 default_gbm    tss_max    0.892 0.0332      5
#> 10 default_maxent boyce_cont 0.802 0.0676      5
#> 11 default_maxent roc_auc    0.988 0.00430     5
#> 12 default_maxent tss_max    0.899 0.0247      5
collect_metrics(lacerta_rep_ens)
#> # A tibble: 18 × 6
#>    rep_id wflow_id       .metric     mean std_err     n
#>    <chr>  <chr>          <chr>      <dbl>   <dbl> <int>
#>  1 rep_01 default_glm    boyce_cont 0.776 0.0518      5
#>  2 rep_01 default_glm    roc_auc    0.963 0.0111      5
#>  3 rep_01 default_glm    tss_max    0.851 0.0270      5
#>  4 rep_01 default_maxent boyce_cont 0.890 0.0184      5
#>  5 rep_01 default_maxent roc_auc    0.989 0.00480     5
#>  6 rep_01 default_maxent tss_max    0.910 0.0316      5
#>  7 rep_02 default_glm    boyce_cont 0.641 0.161       5
#>  8 rep_02 default_glm    roc_auc    0.961 0.0156      5
#>  9 rep_02 default_glm    tss_max    0.861 0.0261      5
#> 10 rep_02 default_maxent boyce_cont 0.878 0.0238      5
#> 11 rep_02 default_maxent roc_auc    0.980 0.00790     5
#> 12 rep_02 default_maxent tss_max    0.880 0.0386      5
#> 13 rep_03 default_glm    boyce_cont 0.467 0.0987      5
#> 14 rep_03 default_glm    roc_auc    0.968 0.0131      5
#> 15 rep_03 default_glm    tss_max    0.869 0.0352      5
#> 16 rep_03 default_maxent boyce_cont 0.787 0.0870      5
#> 17 rep_03 default_maxent roc_auc    0.986 0.00642     5
#> 18 rep_03 default_maxent tss_max    0.921 0.0299      5